requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
aiohttp=3.7.3=pypi_0
async-timeout=3.0.1=pypi_0
attrs=20.3.0=pypi_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=mkl
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.0=pypi_0
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.4=py37h261ae71_0
chardet=3.0.4=pypi_0
conda=4.9.2=py37h06a4308_0
conda-build=3.18.11=py37_0
conda-package-handling=1.7.2=py37h03888b9_0
cryptography=3.3.1=py37h3c74f83_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
cython-bbox=0.1.3=pypi_0
decorator=4.4.2=py_0
fairscale=0.1.4=pypi_0
filelock=3.0.12=py_0
flake8=3.8.4=pypi_0
flake8-import-order=0.18.1=pypi_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.8.5=pypi_0
future=0.18.2=pypi_0
glob2=0.7=py_0
google-auth=1.24.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
grpcio=1.34.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=py_0
imageio=2.9.0=pypi_0
importlib-metadata=3.3.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2020.2=254
ipython=7.15.0=py37_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py37_0
jinja2=2.11.2=py_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lap=0.4.0=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libarchive=3.4.2=h62408e4_0
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
liblief=0.10.1=he6710b0_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libxml2=2.9.10=hb55368b_3
llvmlite=0.36.0=pypi_0
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.3.3=pypi_0
mccabe=0.6.1=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
motmetrics=1.2.0=pypi_0
multidict=5.1.0=pypi_0
ncurses=6.2=he6710b0_1
networkx=2.5.1=pypi_0
ninja=1.10.2=py37hff7bd54_0
numba=0.53.1=pypi_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
omegaconf=2.0.6=pypi_0
opencv-contrib-python=3.4.2.17=pypi_0
opencv-python=3.4.2.17=pypi_0
openssl=1.1.1k=h27cfd23_0
packaging=20.8=pypi_0
pandas=1.2.0=pypi_0
parso=0.8.1=pyhd3eb1b0_0
patchelf=0.12=h2531618_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py37he98fc37_0
pip=20.3.3=py37h06a4308_0
pkginfo=1.6.1=py37h06a4308_0
pluggy=0.13.1=pypi_0
progress=1.5=pypi_0
prompt-toolkit=3.0.8=py_0
protobuf=3.14.0=pypi_0
psutil=5.7.2=py37h7b6447c_0
ptyprocess=0.6.0=pyhd3eb1b0_2
py=1.10.0=pypi_0
py-cpuinfo=7.0.0=pypi_0
py-lief=0.10.1=py37h403a769_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycodestyle=2.6.0=pypi_0
pycosat=0.6.3=py37h27cfd23_0
pycparser=2.20=py_2
pyflakes=2.2.0=pypi_0
pygments=2.7.3=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py37_1
pytest=6.2.1=pypi_0
pytest-benchmark=3.2.3=pypi_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=pypi_0
python-libarchive-c=2.9=py_0
pytorch=1.5.1=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2020.4=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=py37h7b6447c_0
readline=8.0=h7b6447c_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
ripgrep=12.1.1=0
rsa=4.6=pypi_0
ruamel_yaml=0.15.87=py37h7b6447c_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.22.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=51.0.0=py37h06a4308_2
six=1.15.0=py37h06a4308_0
sklearn=0.0=pypi_0
soupsieve=2.1=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.4.8=pypi_0
timm=0.3.4=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
torchaudio=0.5.1=py37
torchvision=0.6.1=py37_cu101
tqdm=4.54.1=pyhd3eb1b0_0
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.2=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xmltodict=0.12.0=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yaml=0.1.7=had09818_2
yarl=1.6.3=pypi_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
